move EditorWidget to separate file
